% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyph.support-en.R
\name{hyph.support.en}
\alias{hyph.support.en}
\title{Language support for English}
\usage{
hyph.support.en()
}
\description{
This function adds support for English to the sylly package. You should not
need to call it manually, as that is done automatically when this package is
loaded.
}
\details{
In particular,
      a new set of hyphenation patterns is being added (see \code{\link{hyph.en}}).
To use the patterns with \code{\link[sylly:hyphen]{hyphen}}, simply use the abbreviation:
\itemize{
  \item {\code{"en"}} {--- English hyphenation patterns (UK)}
  \item {\code{"en.us"}} {--- English hyphenation patterns (US)}
}
}
\seealso{
\code{\link[sylly:hyphen]{hyphen}},
   \code{\link[sylly:set.hyph.support]{set.hyph.support}}
}
